fix(chat): resolve infinite render loop by relocating dnd listener to root
